pyproject.toml
[project]
name = "lda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "chromadb>=1.1.0",
    "fastapi[all]>=0.118.0",
    "google-generativeai>=0.8.5",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "torch>=2.8.0",
    "transformers>=4.56.2",
]
